backend/requirements.txt
Django==5.2.12
djangorestframework>=3.15.2
django-cors-headers>=4.4.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.11
Pillow>=10.0.0
django-jazzmin==3.0.3
